Change write_byte fn signature to take u16 as addr parameter.

+++ src/memory.rs
@@ -1,5 +1,5 @@
 pub trait Memory {
-    fn write_byte(&mut self, addr: usize, byte: u8);
+    fn write_byte(&mut self, addr: u16, byte: u8);
     fn read_byte(&self, addr: u16) -> u8;
 }
 
@@ -17,14 +17,14 @@ impl SimpleMemory {
     pub fn from_rom(rom: &[u8]) -> SimpleMemory {
         let mut mem = SimpleMemory::new();
         for (idx, b) in rom.into_iter().enumerate() {
-            mem.write_byte(idx, *b);
+            mem.write_byte(idx as u16, *b);
         }
         mem
     }
 }
 
 impl Memory for SimpleMemory {
-    fn write_byte(&mut self, addr: usize, byte: u8) {
+    fn write_byte(&mut self, addr: u16, byte: u8) {
         self.inner[addr as usize] = byte;
     }
